+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## Master
+* Disallow bucket without name
 * Fix typos in tests
 * Fix wrong test class name
 * Throw an exception when using a non-existent bucket